+++ src/modules/jobSeeker/dtos/jobSeeker.dto.ts
@@ -1,6 +1,5 @@
 import { Timestamps } from "@types";
 import { UserDto } from "@user/dtos/user.dto";
-import { UserResDto } from "@user/dtos/userRes.dto";
 import { Document } from "mongoose";
 
 export interface JobSeekerDto extends Document, Timestamps {
@@ -9,15 +8,4 @@ export interface JobSeekerDto extends Document, Timestamps {
   lastEducation: string;
   major: string;
   gpa: number;
-}
-
-export interface JobSeekerResponseDto {
-  id: string;
-  user: UserResDto;
-  name: string;
-  lastEducation: string;
-  major: string;
-  gpa: number;
-  createdAt: Date;
-  updatedAt: Date;
 }
